# docs/swagger.yaml
definitions:
  domain.Song:
    properties:
      created_at:
        type: string
      group:
        type: string
      id:
        type: integer
      lyrics:
        type: string
      release_date:
        type: string
      song:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  pkg.ErrorResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Возвращает список песен с возможностью фильтрации и пагинации
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Group
        in: query
        name: group
        type: string
      - description: Song
        in: query
        name: song
        type: string
      - description: Release date
        in: query
        name: release_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Song'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Получение списка песен
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Создает новую песню
      parameters:
      - description: New song
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/domain.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Добавление новой песни
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Удаляет песню по ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Удаление песни
      tags:
      - songs
    get:
      description: Возвращает текст песни по ID c пагинацией по куплетам
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Получение текста песни
      tags:
      - songs
    put:
      consumes:
      - application/json
      description: Обновляет информацию о песне по ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated song
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/domain.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Обновление данных песни
      tags:
      - songs
swagger: "2.0"